Only show select responses
Instead of going through the data and deleting all the values we aren't
using, only use the ones that have a label in the "richardResponses"
column. This way we can keep our data in tact for the time being,
because we haven't really decided yet if we want to use net values or
not on some questions.