Replace bad chars in the data with 0 when parsing CSV for XML
Replace "* " and "- " in the data with "0" because those characters
were breaking stuff, and if the value is 0 that label won't show up
anyway (because it won't have a pixel width, not because it's not added
to the DOM, fix that maybe). Looks like there are checks in the code to
get rid of them but it was still breaking.