Fix stupid javascript float bug
So apparently when you multiply a float by 100 in js, it doesn't really
do it accurately. Ex: 0.28*100 = 28.88888889.
Uhg. Converting to int was truncating it and the right numbers weren't
always showing up in the tooltips. Should have been formatting
converting it to whole numbers outside the XML anyway to avoid having
to do the math in the browser.