Add py script to `.capitalize()` text
Polling company sends it as all CAPS. py/CSVtoLowercase.py will
capitalize first letter of each string and lowercase the rest. Add the
generated file (pollLower.csv) as a sheet to the source spreadsheet, go
through it and capitalize the necessary words, and use that new sheet
as the new source.